+++ internal/cmd/kafka/utils.go
@@ -1,0 +1,10 @@
+package kafka
+
+
+func copyMap(inputMap map[string]string) map[string]string {
+	newMap := make(map[string]string)
+	for key, val := range inputMap {
+		newMap[key] = val
+	}
+	return newMap
+}
